backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic-settings
python-multipart
